First functional version
A functional first version of the add-on.
It registers an HTTP sender listener and sends traces in the Satellite
Message format to http://localhost:9999/1.0/ebpf/traces.

Currently, the add-on is always enabled and the satellite URL cannot be
configured. The functionality to tweak these settings will be added in
subsequent commits.